Moved ConsumeHandle to consume submodule under storage::impls::glommio
